CSV file with all building typologies
Lists all the per default values of building's characteristics, according to the ProRéno data and identification. These values are used in order to complete proxy data when optional questions are not filled by the user.

Colons : Typology, Structure of Housing, Year of Construction ,Number of floors, Structure juxtaposition, Distribution of boiler, Existence of basement, Area of one dwelling, Height of floor, Height of basement, Height of ground, Structure of building, Structure of pavement, Thickness of walls, Existence of elevators, Existence of electrical radiators, Height of Crawl space, Year of renovation