New file to fill missing values when the user doesn't complete all optional questions
Define the complete_user function, taking into argument 'user', the list of user's answer, with mandatory and optional questions. With mandatory questions, complete_user determines the closest typology of building according to ProRéno identification. Then we fill missing values with per default answers per category.

When mandatory values doesn't correspond exactly to our typologies, we establish a weighting ordering the mandatory identifying questions : by order of weight : typology, Year of construction, Number of floors, Structure juxtaposition, Distribution of boiler and Existence of Basement.

Year of construction is the only parameter where we were not able to proxy it with ProRéno data. In case it is not filled by the user, we assume worse situation saying that year of renovation equals year of construction.